started working on an updated and better version of the schema script
popschema became vector populator. virus populator contains an effort to fully automate the process which will be needed for the admin portal for if data is updated.

additionally, I'm working out how to get it to instantiate classes after the dependent tables are filled, and how those tables are filled with their respective data from other tables.