Updated import_data.py, rebuild_all.sh and other associated code to import data via the import_data script.  This should elininate the need for the schema_populator.py script going forward.
There seems to be a discrepancy between whether to use the
Arbovector_database.csv file or the 'main_vector' sheet in
the Arbovirus Excel file.

The virus data comes from the Excel 'main_virus' worksheet.

The current schema_populator script uses the 'main_virus'
worksheet from the Excel file and the CSV file, so that is
what is implemented in the import_data script, but it seems
much of this data is replicated in the 'main_vector' sheet in
the Excel file.